Narrow computer-use e2e trigger
