+++ .github/workflows/computer-e2e.yml
@@ -21,7 +21,6 @@ on:
       - 'src/main/runtime/rpc/errors.ts'
       - 'src/main/runtime/rpc/methods/computer*.ts'
       - 'src/shared/computer-use-*.ts'
-      - 'src/shared/runtime-types.ts'
       - 'tests/e2e/computer-linux.e2e.ts'
       - 'tests/e2e/computer-mac.e2e.ts'
       - 'tests/e2e/computer-mac-safari.e2e.ts'
